Move metrics tracker from plugin reporting to shared
